+++ plantdescriptionengine/src/main/java/eu/arrowhead/core/plantdescriptionengine/services/pde_monitor/dto/PlantDescriptionEntry.java
@@ -43,7 +43,7 @@ public interface PlantDescriptionEntry {
     int id();
     String plantDescription();
     boolean active();
-    List<Integer> include();
+    List<Float> include();
     List<SystemEntry> systems();
     List<Connection> connections();
     Instant createdAt();
